Sparkle 2.3.0
This adds Sparkle 2 to the project add changes the symlink to point to
it. This version has changed the min OS requirement from 10.9 to 10.13,
and as such we keep the old Sparkle 1 around, and during CI, we swap to
it when building legacy binary (regular MacVim has minOS set to 10.13,
while legacy has min OS set to 10.9).