Fix non-CoreText renderer not working well with wide text
This was previously broken by #1287 as I did not update the new
constants' values.

Also, make strikethrough work in it.

The non-CoreText renderer is essentially deprecated and will be removed
in near future, but for now just fix up the straggling bugs since this
is easy to do so.

Fix #1296