Fix code to use getreg() instead
I just missed the fact that getreg() also works and I don't need to hack
to use clip_convert_selection() and the mod on it. This is kind of...
odd, because clip_convert_selection() is essentially doing the same
thing as getreg() with *very* minor differences. You would hope that
would be consolidated and refactored into a single function, but of
course not.

Also cleaned up some comments to make it clear that evaluateExpression
probably would have worked except for the autocmd blocking part.