Enable link-time optimization for Vim in MacVim CI builds
From local profiling, enabling LTO for Vim gives a small but measurable
improvement to performance. One test that I did was to open a really
large Markdown file with vim-markdown (which usually chokes at large
files) installed, and measure how long that takes. With LTO turned on,
usually it gives at least 6-10% performance boost, which seems
significant enough to justify turning it on as we essentially get the
improvement for free (I didn't see similar boosts in other benhcmarking
I did though, so it depends). Slight caveat is that the binary size sees
a small increase (presumably due to inlining) but it's not too much. It
takes more time to build with this turned on though, so only do this in
CI, for the publish builds (we don't do this for the other runs in the
matrix so those runs can finish faster to provide timely feedbacks).

This doesn't change the compilation/linking options for MacVim binary
itself as that doesn't seem to be where performance caps are.