Add redirect and README for old gh-pages branch
This branch is now obsolete, with the new macvim-dev.github.io branch
serving as the canonical website, as it makes more sense to host this on
a root organizational level instead.

Set up a redirect to point to the new site, and add README for
explanation (especially how the appcast/latest.xml file is still
important to keep around for now to support older versions of MacVim).