Fix misc MacVim project warnings in Xcode
Turn on parallel building, to remove the "Building targets in manual
order is deprecated" project warning.

Also, fix configure to only set macOS deployment target to the major
version of macOS when automatically setting it from the client OS
version. E.g. On macOS 13.2 it will set deployment to 13.0. This is
useful because usually we don't want it to be as granular as the minor
version which had caused issues in Homebrew before
(https://github.com/Homebrew/homebrew-core/issues/111693) where they had
to fix on their end, and also 13.2 ends up being too new for Xcode which
only expects to see up to 13.1 as deployment target.

- Note that the logic used to work in OSX 10.X days because the "minor"
  version of X is actually the OS version, whereas in macOS 11/12/13 we
  now have versions like 13.2 where the first number is now the OS
  version. The configure script will now detect whether it's 10.X.Y or
  X.Y (X != 10) and set the correct target correspondingly.